# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Rental API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://dev.book-rental-api/api/v1"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a list of books",
                "operationId": "28b9632495ac5f6f66b2e52001d510e6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Add new book",
                "operationId": "b10f5fe35eefa47e974dd4b333d870b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "synopsis",
                                    "publisher",
                                    "edition",
                                    "page_number",
                                    "isbn",
                                    "language",
                                    "release_date"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Book name",
                                        "type": "string"
                                    },
                                    "synopsis": {
                                        "description": "Book synopsis",
                                        "type": "string"
                                    },
                                    "publisher": {
                                        "description": "Book publisher",
                                        "type": "string"
                                    },
                                    "edition": {
                                        "description": "Book edition",
                                        "type": "string"
                                    },
                                    "page_number": {
                                        "description": "Number of pages in the book",
                                        "type": "integer"
                                    },
                                    "isbn": {
                                        "description": "Book isbn",
                                        "type": "string"
                                    },
                                    "language": {
                                        "description": "Book language",
                                        "type": "string"
                                    },
                                    "release_date": {
                                        "description": "Book release date",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Por uma história-mundo",
                                    "synopsis": "Na era da mundialização, como escrever uma história aberta sobre o mundo...",
                                    "publisher": "Autêntica",
                                    "edition": "1",
                                    "page_number": 200,
                                    "isbn": "9788582176122",
                                    "language": "Português",
                                    "release_date": "25/07/2015"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a book by id",
                "operationId": "e94109dcb4f4d40b58a210b43d94016d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update book",
                "operationId": "7430c1078c3cb5da552b2fd50409f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "synopsis",
                                    "publisher",
                                    "edition",
                                    "page_number",
                                    "isbn",
                                    "language",
                                    "release_date"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Book name",
                                        "type": "string"
                                    },
                                    "synopsis": {
                                        "description": "Book synopsis",
                                        "type": "string"
                                    },
                                    "publisher": {
                                        "description": "Book publisher",
                                        "type": "string"
                                    },
                                    "edition": {
                                        "description": "Book edition",
                                        "type": "string"
                                    },
                                    "page_number": {
                                        "description": "Number of pages in the book",
                                        "type": "integer"
                                    },
                                    "isbn": {
                                        "description": "Book isbn",
                                        "type": "string"
                                    },
                                    "language": {
                                        "description": "Book language",
                                        "type": "string"
                                    },
                                    "release_date": {
                                        "description": "Book release date",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Por uma história-mundo",
                                    "synopsis": "Na era da mundialização, como escrever uma história aberta sobre o mundo...",
                                    "publisher": "Autêntica",
                                    "edition": "1",
                                    "page_number": 200,
                                    "isbn": "9788582176122",
                                    "language": "Português",
                                    "release_date": "25/07/2015"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful update",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book by id",
                "operationId": "41c44a46960ad1a921ba5bfbf16ea7b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}